requirements.txt
Py3Tumblr==0.0.6
python_telegram_bot==5.3.0
pytz>=2016.10
PyVimeo==0.3.9
tzlocal>=1.3
